Fix the provisioning to match the new project layout
As the web-ui will be exported outside of the irma-frontend repository,
we need to separate routes for the API from routes for the interface.
We had to support CORS to have the web-ui to connect to the API.

Moreover, VM name have been changed to use the `host_vars` directory.